+++ c/unreal/Sample/Source/McpBrokerSample/Private/McpBrokerSampleGameMode.cpp
@@ -1,0 +1,9 @@
+#include "McpBrokerSampleGameMode.h"
+#include "McpEchoActor.h"
+#include "Engine/World.h"
+
+void AMcpBrokerSampleGameMode::BeginPlay()
+{
+    Super::BeginPlay();
+    GetWorld()->SpawnActor<AMcpEchoActor>();
+}